+++ src_plugins/ppx_deriving_eq.ml
@@ -5,28 +5,16 @@ open Parsetree
 open Ast_helper
 open Ast_convenience
 
-module StringSet = Ppx_deriving.StringSet
-
 let deriver = "eq"
 let raise_errorf = Ppx_deriving.raise_errorf
 
-type eq_options =
-  {
-    allow_std_type_shadowing: bool;
-  }
-
-let default_eq_options =
-  {
-    allow_std_type_shadowing= false;
-  }
-
 let parse_options options =
-  let option_parser acc (name, expr) =
+  options |> List.iter (fun (name, expr) ->
     match name with
-    | "allow_std_type_shadowing" -> { allow_std_type_shadowing = true }
-    | _ ->
-      raise_errorf ~loc:expr.pexp_loc "%s does not support option %s" deriver name in
-  List.fold_left option_parser default_eq_options options
+    | _ -> raise_errorf ~loc:expr.pexp_loc "%s does not support option %s" deriver name)
+
+let attr_nobuiltin attrs =
+  Ppx_deriving.(attrs |> attr ~deriver "nobuiltin" |> Arg.get_flag ~deriver)
 
 let attr_equal attrs =
   Ppx_deriving.(attrs |> attr ~deriver "equal" |> Arg.(get_attr ~deriver expr))
@@ -38,63 +26,72 @@ let pattn side typs =
   List.mapi (fun i _ -> pvar (argn side i)) typs
 
 let core_type_of_decl ~options ~path type_decl =
-  ignore (parse_options options);
+  parse_options options;
   let typ = Ppx_deriving.core_type_of_type_decl type_decl in
   Ppx_deriving.poly_arrow_of_type_decl
-    (fun var -> [%type: [%t var] -> [%t var] -> bool])
+    (fun var -> [%type: [%t var] -> [%t var] -> Ppx_deriving_runtime.bool])
     type_decl
-    [%type: [%t typ] -> [%t typ] -> bool]
+    [%type: [%t typ] -> [%t typ] -> Ppx_deriving_runtime.bool]
 
 let sig_of_type ~options ~path type_decl =
-  ignore (parse_options options);
+  parse_options options;
   [Sig.value (Val.mk (mknoloc (Ppx_deriving.mangle_type_decl (`Prefix "equal") type_decl))
              (core_type_of_decl ~options ~path type_decl))]
 
-let rec exprsn group_def typs =
+let rec exprsn quoter typs =
   typs |> List.mapi (fun i typ ->
-    app (expr_of_typ group_def typ) [evar (argn `lhs i); evar (argn `rhs i)])
+    app (expr_of_typ quoter typ) [evar (argn `lhs i); evar (argn `rhs i)])
 
-and expr_of_typ group_def typ =
+and expr_of_typ quoter typ =
+  let expr_of_typ = expr_of_typ quoter in
   match attr_equal typ.ptyp_attributes with
-  | Some fn -> fn
+  | Some fn -> Ppx_deriving.quote quoter fn
   | None ->
     match typ with
-    | { ptyp_desc = Ptyp_constr ({ txt = (Lident id as lid) }, args) } 
-        when StringSet.mem id group_def ->
-      let equal_fn = Exp.ident (mknoloc (Ppx_deriving.mangle_lid (`Prefix "equal") lid)) in
-      app equal_fn (List.map (expr_of_typ group_def) args)
-    | [%type: _] | [%type: unit] -> [%expr fun _ _ -> true]
-    | [%type: int] | [%type: int32] | [%type: Int32.t]
-    | [%type: int64] | [%type: Int64.t] | [%type: nativeint] | [%type: Nativeint.t]
-    | [%type: float] | [%type: bool] | [%type: char] | [%type: string] |
-      [%type: String.t] | [%type: bytes] ->
-      [%expr (fun (a:[%t typ]) b -> a = b)]
-    | [%type: [%t? typ] ref]   -> [%expr fun a b -> [%e expr_of_typ group_def typ] !a !b]
-    | [%type: [%t? typ] list]  ->
-      [%expr
-        let rec loop x y =
+    | [%type: _] -> [%expr fun _ _ -> true]
+    | { ptyp_desc = Ptyp_constr _ } ->
+      let builtin = not (attr_nobuiltin typ.ptyp_attributes) in
+      begin match builtin, typ with
+      | true, [%type: unit] ->
+        [%expr fun _ _ -> true]
+      | true, ([%type: int] | [%type: int32] | [%type: Int32.t] |
+               [%type: int64] | [%type: Int64.t] | [%type: nativeint] |
+               [%type: Nativeint.t] | [%type: float] | [%type: bool] |
+               [%type: char] | [%type: string] | [%type: bytes]) ->
+        [%expr (fun (a:[%t typ]) b -> a = b)]
+      | true, [%type: [%t? typ] ref] ->
+        [%expr fun a b -> [%e expr_of_typ typ] !a !b]
+      | true, [%type: [%t? typ] list] ->
+        [%expr
+          let rec loop x y =
+            match x, y with
+            | [], [] -> true
+            | a :: x, b :: y -> [%e expr_of_typ typ] a b && loop x y
+            | _ -> false
+          in (fun x y -> loop x y)]
+      | true, [%type: [%t? typ] array] ->
+        [%expr fun x y ->
+          let rec loop i =
+            (i = Array.length x || [%e expr_of_typ typ] x.(i) y.(i)) && loop (i + 1)
+          in Array.length x = Array.length y && loop 0]
+      | true, [%type: [%t? typ] option] ->
+        [%expr fun x y ->
           match x, y with
-          | [], [] -> true
-          | a :: x, b :: y -> [%e expr_of_typ group_def typ] a b && loop x y
-          | _ -> false
-        in (fun x y -> loop x y)]
-    | [%type: [%t? typ] array] ->
-      [%expr fun x y ->
-        let rec loop i =
-          (i = Array.length x || [%e expr_of_typ group_def typ] x.(i) y.(i)) && loop (i + 1)
-        in Array.length x = Array.length y && loop 0]
-    | [%type: [%t? typ] option] ->
-      [%expr fun x y ->
-        match x, y with
-        | None, None -> true
-        | Some a, Some b -> [%e expr_of_typ group_def typ] a b
-        | _ -> false]
-    | { ptyp_desc = Ptyp_constr ({ txt = lid }, args) } ->
-      let equal_fn = Exp.ident (mknoloc (Ppx_deriving.mangle_lid (`Prefix "equal") lid)) in
-      app equal_fn (List.map (expr_of_typ group_def) args)
+          | None, None -> true
+          | Some a, Some b -> [%e expr_of_typ typ] a b
+          | _ -> false]
+      | true, ([%type: [%t? typ] lazy_t] | [%type: [%t? typ] Lazy.t]) ->
+        [%expr fun (lazy x) (lazy y) -> [%e expr_of_typ typ] x y]
+      | _, { ptyp_desc = Ptyp_constr ({ txt = lid }, args) } ->
+        let equal_fn = Exp.ident (mknoloc (Ppx_deriving.mangle_lid (`Prefix "equal") lid)) in
+        let fwd = app (Ppx_deriving.quote quoter equal_fn) (List.map expr_of_typ args) in
+        (* eta-expansion is necessary for recursive groups *)
+        [%expr fun x -> [%e fwd] x]
+      | _ -> assert false
+      end
     | { ptyp_desc = Ptyp_tuple typs } ->
       [%expr fun [%p ptuple (pattn `lhs typs)] [%p ptuple (pattn `rhs typs)] ->
-        [%e exprsn group_def typs |> Ppx_deriving.(fold_exprs (binop_reduce [%expr (&&)]))]]
+        [%e exprsn quoter typs |> Ppx_deriving.(fold_exprs (binop_reduce [%expr (&&)]))]]
     | { ptyp_desc = Ptyp_variant (fields, _, _); ptyp_loc } ->
       let cases =
         (fields |> List.map (fun field ->
@@ -104,10 +101,10 @@ and expr_of_typ group_def typ =
             Exp.case (pdup (fun _ -> Pat.variant label None)) [%expr true]
           | Rtag (label, _, false, [typ]) ->
             Exp.case (pdup (fun var -> Pat.variant label (Some (pvar var))))
-                     (app (expr_of_typ group_def typ) [evar "lhs"; evar "rhs"])
+                     (app (expr_of_typ typ) [evar "lhs"; evar "rhs"])
           | Rinherit ({ ptyp_desc = Ptyp_constr (tname, _) } as typ) ->
             Exp.case (pdup (fun var -> Pat.alias (Pat.type_ tname) (mknoloc var)))
-                     (app (expr_of_typ group_def typ) [evar "lhs"; evar "rhs"])
+                     (app (expr_of_typ typ) [evar "lhs"; evar "rhs"])
           | _ ->
             raise_errorf ~loc:ptyp_loc "%s cannot be derived for %s"
                          deriver (Ppx_deriving.string_of_core_type typ))) @
@@ -115,16 +112,17 @@ and expr_of_typ group_def typ =
       in
       [%expr fun lhs rhs -> [%e Exp.match_ [%expr lhs, rhs] cases]]
     | { ptyp_desc = Ptyp_var name } -> evar ("poly_"^name)
-    | { ptyp_desc = Ptyp_alias (typ, _) } -> expr_of_typ group_def typ
+    | { ptyp_desc = Ptyp_alias (typ, _) } -> expr_of_typ typ
     | { ptyp_loc } ->
       raise_errorf ~loc:ptyp_loc "%s cannot be derived for %s"
                    deriver (Ppx_deriving.string_of_core_type typ)
 
-let str_of_type ~options ~path group_def ({ ptype_loc = loc } as type_decl) =
-  ignore (parse_options options);
+let str_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
+  parse_options options;
+  let quoter = Ppx_deriving.create_quoter () in
   let comparator =
     match type_decl.ptype_kind, type_decl.ptype_manifest with
-    | Ptype_abstract, Some manifest -> expr_of_typ group_def manifest
+    | Ptype_abstract, Some manifest -> expr_of_typ quoter manifest
     | Ptype_variant constrs, _ ->
       let wildcard = match constrs with
         | [] | [_] -> []
@@ -134,7 +132,7 @@ let str_of_type ~options ~path group_def ({ ptype_loc = loc } as type_decl) =
          List.map
            (function
              | { pcd_name = { txt = name }; pcd_args = Pcstr_tuple typs } ->
-               exprsn group_def typs |>
+               exprsn quoter typs |>
                Ppx_deriving.(fold_exprs ~unit:[%expr true] (binop_reduce [%expr (&&)])) |>
                Exp.case (ptuple [pconstr name (pattn `lhs typs);
                                  pconstr name (pattn `rhs typs)])
@@ -150,7 +148,7 @@ let str_of_type ~options ~path group_def ({ ptype_loc = loc } as type_decl) =
           let attrs =  pld_type.ptyp_attributes @ pld_attributes in
           let pld_type = {pld_type with ptyp_attributes=attrs} in
           let field obj = Exp.field obj (mknoloc (Lident name)) in
-          app (expr_of_typ group_def pld_type) [field (evar "lhs"); field (evar "rhs")])
+          app (expr_of_typ quoter pld_type) [field (evar "lhs"); field (evar "rhs")])
       in
       [%expr fun lhs rhs -> [%e exprs |> Ppx_deriving.(fold_exprs (binop_reduce [%expr (&&)]))]]
     | Ptype_abstract, None ->
@@ -162,30 +160,16 @@ let str_of_type ~options ~path group_def ({ ptype_loc = loc } as type_decl) =
   let out_type =
     Ppx_deriving.strong_type_of_type @@
       core_type_of_decl  ~options ~path type_decl in
-  let eq_var = 
+  let eq_var =
     pvar (Ppx_deriving.mangle_type_decl (`Prefix "equal") type_decl) in
-  [Vb.mk (Pat.constraint_ eq_var out_type) (polymorphize comparator)]
-
-let type_decl_str ~options ~path type_decls =
-  let opts = parse_options options in
-  let typename_set =
-    Ppx_deriving.extract_typename_of_type_group 
-      deriver
-      ~allow_shadowing:opts.allow_std_type_shadowing
-      type_decls in
-  let here_loc = (List.hd type_decls).ptype_loc in
-  if StringSet.mem "bool" typename_set then
-    raise_errorf
-      ~loc:here_loc
-      "%s can't derivate types when shadowing bool (even with option)" deriver;
-  let code =
-    List.map (str_of_type ~options ~path typename_set) type_decls in
-  [Str.value Recursive (List.concat code)]
+  [Vb.mk (Pat.constraint_ eq_var out_type)
+         (Ppx_deriving.sanitize ~quoter (polymorphize comparator))]
 
 let () =
-  Ppx_deriving.(register (create deriver
-    ~core_type:(expr_of_typ StringSet.empty)
-    ~type_decl_str: type_decl_str
+  Ppx_deriving.(register (create "eq"
+    ~core_type: (Ppx_deriving.with_quoter expr_of_typ)
+    ~type_decl_str: (fun ~options ~path type_decls ->
+       [Str.value Recursive (List.concat (List.map (str_of_type ~options ~path) type_decls))])
     ~type_decl_sig: (fun ~options ~path type_decls ->
        List.concat (List.map (sig_of_type ~options ~path) type_decls))
     ()