+++ src_plugins/ppx_deriving_ord.ml
@@ -5,28 +5,16 @@ open Parsetree
 open Ast_helper
 open Ast_convenience
 
-module StringSet = Ppx_deriving.StringSet
-
 let deriver = "ord"
 let raise_errorf = Ppx_deriving.raise_errorf
 
-type ord_options =
-  {
-    allow_std_type_shadowing: bool;
-  }
-
-let default_ord_options =
-  {
-    allow_std_type_shadowing = false;
-  }
-
 let parse_options options =
-  let option_parser acc (name, expr) =
+  options |> List.iter (fun (name, expr) ->
     match name with
-    | "allow_std_type_shadowing" -> { allow_std_type_shadowing = true }
-    | _ ->
-      raise_errorf ~loc:expr.pexp_loc "%s does not support option %s" deriver name in
-  List.fold_left option_parser default_ord_options options
+    | _ -> raise_errorf ~loc:expr.pexp_loc "%s does not support option %s" deriver name)
+
+let attr_nobuiltin attrs =
+  Ppx_deriving.(attrs |> attr ~deriver "nobuiltin" |> Arg.get_flag ~deriver)
 
 let attr_compare attrs =
   Ppx_deriving.(attrs |> attr ~deriver "compare" |> Arg.(get_attr ~deriver expr))
@@ -46,35 +34,33 @@ let wildcard_case int_cases =
     let to_int = [%e Exp.function_ int_cases] in
     Pervasives.compare (to_int lhs) (to_int rhs)]
 
-(* deactivate warning 4 in code that uses [wildcard_case] *)
-let warning_minus_4 =
-  Ppx_deriving.attr_warning [%expr "-4"]
-
 let pattn side typs =
   List.mapi (fun i _ -> pvar (argn side i)) typs
 
-let rec exprsn group_def typs =
+let rec exprsn quoter typs =
   typs |> List.mapi (fun i typ ->
-    app (expr_of_typ group_def typ) [evar (argn `lhs i); evar (argn `rhs i)])
+    app (expr_of_typ quoter typ) [evar (argn `lhs i); evar (argn `rhs i)])
 
-and expr_of_typ group_def typ =
-  let rec aux typ =
-    match attr_compare typ.ptyp_attributes with
-    | Some fn -> fn
-    | None ->
-      match typ with
-      | { ptyp_desc = Ptyp_constr ({ txt = (Lident id as lid) }, args) } when
-            StringSet.mem id group_def ->
-        let compare_fn = Exp.ident (mknoloc (Ppx_deriving.mangle_lid (`Prefix "compare") lid)) in
-        app compare_fn (List.map aux args)
-      | [%type: _] | [%type: unit] -> [%expr fun _ _ -> 0]
-      | [%type: int] | [%type: int32] | [%type: Int32.t]
-      | [%type: int64] | [%type: Int64.t] | [%type: nativeint] | [%type: Nativeint.t]
-      | [%type: float] | [%type: bool] | [%type: char] | [%type: string] | [%type: String.t] |
-        [%type: bytes] ->
+and expr_of_typ quoter typ =
+  let expr_of_typ = expr_of_typ quoter in
+  match attr_compare typ.ptyp_attributes with
+  | Some fn -> Ppx_deriving.quote quoter fn
+  | None ->
+    match typ with
+    | [%type: _] -> [%expr fun _ _ -> 0]
+    | { ptyp_desc = Ptyp_constr _ } ->
+      let builtin = not (attr_nobuiltin typ.ptyp_attributes) in
+      begin match builtin, typ with
+      | true, ([%type: _] | [%type: unit]) ->
+        [%expr fun _ _ -> 0]
+      | true, ([%type: int] | [%type: int32] | [%type: Int32.t]
+              | [%type: int64] | [%type: Int64.t] | [%type: nativeint]
+              | [%type: Nativeint.t] | [%type: float] | [%type: bool]
+              | [%type: char] | [%type: string] | [%type: bytes]) ->
         [%expr Pervasives.compare]
-      | [%type: [%t? typ] ref]   -> [%expr fun a b -> [%e aux typ] !a !b]
-      | [%type: [%t? typ] list]  ->
+      | true, [%type: [%t? typ] ref] ->
+        [%expr fun a b -> [%e expr_of_typ typ] !a !b]
+      | true, [%type: [%t? typ] list]  ->
         [%expr
           let rec loop x y =
             match x, y with
@@ -82,83 +68,89 @@ and expr_of_typ group_def typ =
             | [], _ -> -1
             | _, [] -> 1
             | a :: x, b :: y ->
-              [%e compare_reduce [%expr loop x y] [%expr [%e aux typ] a b]]
+              [%e compare_reduce [%expr loop x y] [%expr [%e expr_of_typ typ] a b]]
           in (fun x y -> loop x y)]
-      | [%type: [%t? typ] array] ->
+      | true, [%type: [%t? typ] array] ->
         [%expr fun x y ->
           let rec loop i =
             if i = Array.length x then 0
             else [%e compare_reduce [%expr loop (i + 1)]
-                                    [%expr [%e aux typ] x.(i) y.(i)]]
+                                    [%expr [%e expr_of_typ typ] x.(i) y.(i)]]
           in
           [%e compare_reduce [%expr loop 0]
                              [%expr Pervasives.compare (Array.length x) (Array.length y)]]]
-      | [%type: [%t? typ] option] ->
+      | true, [%type: [%t? typ] option] ->
         [%expr fun x y ->
           match x, y with
           | None, None -> 0
-          | Some a, Some b -> [%e aux typ] a b
+          | Some a, Some b -> [%e expr_of_typ typ] a b
           | None, Some _ -> -1
           | Some _, None -> 1]
-      | { ptyp_desc = Ptyp_constr ({ txt = lid }, args) } ->
+      | true, ([%type: [%t? typ] lazy_t] | [%type: [%t? typ] Lazy.t]) ->
+        [%expr fun (lazy x) (lazy y) -> [%e expr_of_typ typ] x y]
+      | _, { ptyp_desc = Ptyp_constr ({ txt = lid }, args) } ->
         let compare_fn = Exp.ident (mknoloc (Ppx_deriving.mangle_lid (`Prefix "compare") lid)) in
-        app compare_fn (List.map aux args)
-      | { ptyp_desc = Ptyp_tuple typs } ->
-        [%expr fun [%p ptuple (pattn `lhs typs)] [%p ptuple (pattn `rhs typs)] ->
-          [%e exprsn group_def typs |> List.rev |> reduce_compare]]
-      | { ptyp_desc = Ptyp_variant (fields, _, _); ptyp_loc } ->
-        let cases =
-          fields |> List.map (fun field ->
-            let pdup f = ptuple [f "lhs"; f "rhs"] in
-            match field with
-            | Rtag (label, _, true (*empty*), []) ->
-              Exp.case (pdup (fun _ -> Pat.variant label None)) [%expr 0]
-            | Rtag (label, _, false, [typ]) ->
-              Exp.case (pdup (fun var -> Pat.variant label (Some (pvar var))))
-                       (app (aux typ) [evar "lhs"; evar "rhs"])
-            | Rinherit ({ ptyp_desc = Ptyp_constr (tname, _) } as typ) ->
-              Exp.case (pdup (fun var -> Pat.alias (Pat.type_ tname) (mknoloc var)))
-                       (app (aux typ) [evar "lhs"; evar "rhs"])
-            | _ ->
-              raise_errorf ~loc:ptyp_loc "%s cannot be derived for %s"
-                           deriver (Ppx_deriving.string_of_core_type typ))
-        in
-        let int_cases =
-          fields |> List.mapi (fun i field ->
-            match field with
-            | Rtag (label, _, true (*empty*), []) ->
-              Exp.case (Pat.variant label None) (int i)
-            | Rtag (label, _, false, [typ]) ->
-              Exp.case (Pat.variant label (Some [%pat? _])) (int i)
-            | Rinherit { ptyp_desc = Ptyp_constr (tname, []) } ->
-              Exp.case (Pat.type_ tname) (int i)
-            | _ -> assert false)
-        in
-        [%expr fun lhs rhs ->
-          [%e Exp.match_ ~attrs:[warning_minus_4] [%expr lhs, rhs] (cases @ [wildcard_case int_cases])]]
-      | { ptyp_desc = Ptyp_var name } -> evar ("poly_"^name)
-      | { ptyp_desc = Ptyp_alias (typ, _) } -> aux typ
-      | { ptyp_loc } ->
-        raise_errorf ~loc:ptyp_loc "%s cannot be derived for %s"
-                     deriver (Ppx_deriving.string_of_core_type typ) in
-  aux typ
+        let fwd = app (Ppx_deriving.quote quoter compare_fn) (List.map expr_of_typ args) in
+        (* eta-expansion is necessary for recursive groups *)
+        [%expr fun x -> [%e fwd] x]
+      | _ -> assert false
+      end
+    | { ptyp_desc = Ptyp_tuple typs } ->
+      [%expr fun [%p ptuple (pattn `lhs typs)] [%p ptuple (pattn `rhs typs)] ->
+        [%e exprsn quoter typs |> List.rev |> reduce_compare]]
+    | { ptyp_desc = Ptyp_variant (fields, _, _); ptyp_loc } ->
+      let cases =
+        fields |> List.map (fun field ->
+          let pdup f = ptuple [f "lhs"; f "rhs"] in
+          match field with
+          | Rtag (label, _, true (*empty*), []) ->
+            Exp.case (pdup (fun _ -> Pat.variant label None)) [%expr 0]
+          | Rtag (label, _, false, [typ]) ->
+            Exp.case (pdup (fun var -> Pat.variant label (Some (pvar var))))
+                     (app (expr_of_typ typ) [evar "lhs"; evar "rhs"])
+          | Rinherit ({ ptyp_desc = Ptyp_constr (tname, _) } as typ) ->
+            Exp.case (pdup (fun var -> Pat.alias (Pat.type_ tname) (mknoloc var)))
+                     (app (expr_of_typ typ) [evar "lhs"; evar "rhs"])
+          | _ ->
+            raise_errorf ~loc:ptyp_loc "%s cannot be derived for %s"
+                         deriver (Ppx_deriving.string_of_core_type typ))
+      in
+      let int_cases =
+        fields |> List.mapi (fun i field ->
+          match field with
+          | Rtag (label, _, true (*empty*), []) ->
+            Exp.case (Pat.variant label None) (int i)
+          | Rtag (label, _, false, [typ]) ->
+            Exp.case (Pat.variant label (Some [%pat? _])) (int i)
+          | Rinherit { ptyp_desc = Ptyp_constr (tname, []) } ->
+            Exp.case (Pat.type_ tname) (int i)
+          | _ -> assert false)
+      in
+      [%expr fun lhs rhs ->
+        [%e Exp.match_ [%expr lhs, rhs] (cases @ [wildcard_case int_cases])]]
+    | { ptyp_desc = Ptyp_var name } -> evar ("poly_"^name)
+    | { ptyp_desc = Ptyp_alias (typ, _) } -> expr_of_typ typ
+    | { ptyp_loc } ->
+      raise_errorf ~loc:ptyp_loc "%s cannot be derived for %s"
+                   deriver (Ppx_deriving.string_of_core_type typ)
 
 let core_type_of_decl ~options ~path type_decl =
-  ignore (parse_options options);
+  parse_options options;
   let typ = Ppx_deriving.core_type_of_type_decl type_decl in
   let polymorphize = Ppx_deriving.poly_arrow_of_type_decl
-          (fun var -> [%type: [%t var] -> [%t var] -> int]) type_decl in
-  (polymorphize [%type: [%t typ] -> [%t typ] -> int])
+          (fun var -> [%type: [%t var] -> [%t var] -> Ppx_deriving_runtime.int]) type_decl in
+  (polymorphize [%type: [%t typ] -> [%t typ] -> Ppx_deriving_runtime.int])
 
 let sig_of_type ~options ~path type_decl =
   [Sig.value (Val.mk (mknoloc (Ppx_deriving.mangle_type_decl (`Prefix "compare") type_decl))
              (core_type_of_decl ~options ~path type_decl))]
 
-let str_of_type ~options ~path group_def ({ ptype_loc = loc } as type_decl) =
-  ignore (parse_options options);
+let str_of_type ~options ~path ({ ptype_loc = loc } as type_decl) =
+  parse_options options;
+  let quoter = Ppx_deriving.create_quoter () in
   let comparator =
     match type_decl.ptype_kind, type_decl.ptype_manifest with
-    | Ptype_abstract, Some manifest -> expr_of_typ group_def manifest
+    | Ptype_abstract, Some manifest -> expr_of_typ quoter manifest
     | Ptype_variant constrs, _ ->
       let int_cases =
         constrs |>
@@ -177,24 +169,21 @@ let str_of_type ~options ~path group_def ({ ptype_loc = loc } as type_decl) =
         List.map
           (function
             | { pcd_name = { txt = name }; pcd_args = Pcstr_tuple typs } ->
-              exprsn group_def typs |> List.rev |> reduce_compare |>
+              exprsn quoter typs |> List.rev |> reduce_compare |>
               Exp.case (ptuple [pconstr name (pattn `lhs typs);
                                 pconstr name (pattn `rhs typs)])
             | { pcd_args = Pcstr_record _ } ->
               raise_errorf ~loc "%s currently doesn't support inline records" deriver) in
       (* if the type as only one constructor, generating wildcard yield a warning. *)
-      let wildcard = match constrs with
-        | [] | [_] -> []
-        | _ :: _ :: _ -> [wildcard_case int_cases] in
       [%expr fun lhs rhs ->
-        [%e Exp.match_ ~attrs:[warning_minus_4] [%expr lhs, rhs] (cases @ wildcard)]]
+        [%e Exp.match_ [%expr lhs, rhs] (cases @ [wildcard_case int_cases])]]
     | Ptype_record labels, _ ->
       let exprs =
         labels |> List.map (fun { pld_name = { txt = name }; pld_type; pld_attributes } ->
           let attrs = pld_attributes @ pld_type.ptyp_attributes in
           let pld_type = {pld_type with ptyp_attributes=attrs} in
           let field obj = Exp.field obj (mknoloc (Lident name)) in
-          app (expr_of_typ group_def pld_type) [field (evar "lhs"); field (evar "rhs")])
+          app (expr_of_typ quoter pld_type) [field (evar "lhs"); field (evar "rhs")])
       in
       [%expr fun lhs rhs -> [%e reduce_compare exprs]]
     | Ptype_abstract, None ->
@@ -209,28 +198,13 @@ let str_of_type ~options ~path group_def ({ ptype_loc = loc } as type_decl) =
   let out_var =
     pvar (Ppx_deriving.mangle_type_decl (`Prefix "compare") type_decl) in
   [Vb.mk (Pat.constraint_ out_var out_type)
-         (polymorphize comparator)]
-
-let type_decl_str ~options ~path type_decls =
-  let opts = parse_options options in
-  let typename_set =
-    Ppx_deriving.extract_typename_of_type_group
-      deriver
-      ~allow_shadowing:opts.allow_std_type_shadowing
-      type_decls in
-  let here_loc = (List.hd type_decls).ptype_loc in
-  if StringSet.mem "int" typename_set then
-    raise_errorf
-      ~loc:here_loc
-      "%s can't derivate types when shadowing int (even with option)" deriver;
-  let code =
-    List.map (str_of_type ~options ~path typename_set) type_decls in
-  [Str.value Recursive (List.concat code)]
+         (Ppx_deriving.sanitize ~quoter (polymorphize comparator))]
 
 let () =
   Ppx_deriving.(register (create deriver
-    ~core_type: (expr_of_typ StringSet.empty)
-    ~type_decl_str: type_decl_str
+    ~core_type: (Ppx_deriving.with_quoter expr_of_typ)
+    ~type_decl_str: (fun ~options ~path type_decls ->
+      [Str.value Recursive (List.concat (List.map (str_of_type ~options ~path) type_decls))])
     ~type_decl_sig: (fun ~options ~path type_decls ->
       List.concat (List.map (sig_of_type ~options ~path) type_decls))
     ()